Перенес логику отрисовки точек маршрута в контроллер
